Add files via upload
HomeScreen: Added a Remaining Balance that subtracts what each person puts in to determine how much is left to pay.
PersonScreen: Wasted a few hours figuring out how route works to try and pass on the inputs to homescreen but it keeps giving issues. Tried using navigation but that did not work. Tried making an index for Person to create multiple pages, tried having a prefixed amount but that didint work either.
Still have a bunch of issues that I can't seem to fix. Next Person doesnt even load. 
in this version PersonScreen is broken though so you cant even get there